pyproject.toml
[project]
name = "enrich-me"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastparquet>=2024.5.0",
    "ipython>=8.28.0",
    "ipywidgets>=8.1.5",
    "jquants-api-client>=1.7.2",
    "lightgbm>=4.5.0",
    "matplotlib>=3.9.2",
    "nodriver>=0.37",
    "numpy>=1.26.4",
    "pandas>=2.2.3",
    "plotly>=5.24.1",
    "pyarrow>=17.0.0",
    "python-dotenv>=1.0.1",
    "quantstats>=0.0.62",
    "scikit-learn>=1.5.2",
    "scipy>=1.14.1",
]
